# CMS-Web/Metadata/apiDefinition.swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"CMS_API"},"host":"localhost:62721","schemes":["http"],"paths":{"/api/Account/UserInfo":{"get":{"tags":["Account"],"operationId":"Account_GetUserInfo","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserInfoViewModel"}}}}},"/api/Account/Logout":{"post":{"tags":["Account"],"operationId":"Account_Logout","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/Account/ManageInfo":{"get":{"tags":["Account"],"operationId":"Account_GetManageInfo","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"returnUrl","in":"query","required":true,"type":"string"},{"name":"generateState","in":"query","required":false,"type":"boolean"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ManageInfoViewModel"}}}}},"/api/Account/ChangePassword":{"post":{"tags":["Account"],"operationId":"Account_ChangePassword","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"model","in":"body","required":true,"schema":{"$ref":"#/definitions/ChangePasswordBindingModel"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/Account/SetPassword":{"post":{"tags":["Account"],"operationId":"Account_SetPassword","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"model","in":"body","required":true,"schema":{"$ref":"#/definitions/SetPasswordBindingModel"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/Account/AddExternalLogin":{"post":{"tags":["Account"],"operationId":"Account_AddExternalLogin","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"model","in":"body","required":true,"schema":{"$ref":"#/definitions/AddExternalLoginBindingModel"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/Account/RemoveLogin":{"post":{"tags":["Account"],"operationId":"Account_RemoveLogin","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"model","in":"body","required":true,"schema":{"$ref":"#/definitions/RemoveLoginBindingModel"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/Account/ExternalLogin":{"get":{"tags":["Account"],"operationId":"Account_GetExternalLogin","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"provider","in":"query","required":true,"type":"string"},{"name":"error","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/Account/ExternalLogins":{"get":{"tags":["Account"],"operationId":"Account_GetExternalLogins","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"returnUrl","in":"query","required":true,"type":"string"},{"name":"generateState","in":"query","required":false,"type":"boolean"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/ExternalLoginViewModel"}}}}}},"/api/Account/Register":{"post":{"tags":["Account"],"operationId":"Account_Register","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"model","in":"body","required":true,"schema":{"$ref":"#/definitions/RegisterBindingModel"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/Account/RegisterExternal":{"post":{"tags":["Account"],"operationId":"Account_RegisterExternal","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"model","in":"body","required":true,"schema":{"$ref":"#/definitions/RegisterExternalBindingModel"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/ImportCounries":{"post":{"tags":["Countries"],"operationId":"Countries_ImportCountries","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/Countries":{"get":{"tags":["Countries"],"operationId":"Countries_Get","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"string"}}}}},"post":{"tags":["Countries"],"operationId":"Countries_Post","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":[],"parameters":[{"name":"value","in":"body","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"}}}},"/api/Countries/{id}":{"get":{"tags":["Countries"],"operationId":"Countries_Get","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}}}},"put":{"tags":["Countries"],"operationId":"Countries_Put","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"value","in":"body","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"}}},"delete":{"tags":["Countries"],"operationId":"Countries_Delete","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"204":{"description":"No Content"}}}},"/api/Values":{"get":{"tags":["Values"],"operationId":"Values_Get","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"string"}}}}},"post":{"tags":["Values"],"operationId":"Values_Post","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":[],"parameters":[{"name":"value","in":"body","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"}}}},"/api/Values/{id}":{"get":{"tags":["Values"],"operationId":"Values_Get","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}}}},"put":{"tags":["Values"],"operationId":"Values_Put","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"value","in":"body","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"}}},"delete":{"tags":["Values"],"operationId":"Values_Delete","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"204":{"description":"No Content"}}}}},"definitions":{"UserInfoViewModel":{"type":"object","properties":{"Email":{"type":"string"},"HasRegistered":{"type":"boolean"},"LoginProvider":{"type":"string"}}},"ManageInfoViewModel":{"type":"object","properties":{"LocalLoginProvider":{"type":"string"},"Email":{"type":"string"},"Logins":{"type":"array","items":{"$ref":"#/definitions/UserLoginInfoViewModel"}},"ExternalLoginProviders":{"type":"array","items":{"$ref":"#/definitions/ExternalLoginViewModel"}}}},"UserLoginInfoViewModel":{"type":"object","properties":{"LoginProvider":{"type":"string"},"ProviderKey":{"type":"string"}}},"ExternalLoginViewModel":{"type":"object","properties":{"Name":{"type":"string"},"Url":{"type":"string"},"State":{"type":"string"}}},"ChangePasswordBindingModel":{"required":["OldPassword","NewPassword"],"type":"object","properties":{"OldPassword":{"type":"string"},"NewPassword":{"maxLength":100,"minLength":6,"type":"string"},"ConfirmPassword":{"type":"string"}}},"SetPasswordBindingModel":{"required":["NewPassword"],"type":"object","properties":{"NewPassword":{"maxLength":100,"minLength":6,"type":"string"},"ConfirmPassword":{"type":"string"}}},"AddExternalLoginBindingModel":{"required":["ExternalAccessToken"],"type":"object","properties":{"ExternalAccessToken":{"type":"string"}}},"RemoveLoginBindingModel":{"required":["LoginProvider","ProviderKey"],"type":"object","properties":{"LoginProvider":{"type":"string"},"ProviderKey":{"type":"string"}}},"RegisterBindingModel":{"required":["Email","Password"],"type":"object","properties":{"Email":{"type":"string"},"Password":{"maxLength":100,"minLength":6,"type":"string"},"ConfirmPassword":{"type":"string"}}},"RegisterExternalBindingModel":{"required":["Email"],"type":"object","properties":{"Email":{"type":"string"}}}}}